OCR/requirements.txt
flask 
os
pytesseract
numpy
pdf2image
google.generativeai
opencv-python 